fix: fixed lint issues in regards to mixing tsx and ts exports on same file
